refactor: move classes to more appropriate packages
