+++ armadillo/src/main/java/org/molgenis/armadillo/storage/FileDownloader.java
@@ -1,4 +1,4 @@
-package org.molgenis.armadillo.controller;
+package org.molgenis.armadillo.storage;
 
 import java.io.BufferedInputStream;
 import java.io.FileOutputStream;
